refactor searchInput component
