+++ src/components/SearchInput/SearchInput.scss
